Adding tower's shooting script and add behaviour for the enemy when they got shot
I have added a script where tower can scan its surrounding and shoot if its an enemy, also they can rotate to the enemy position.

The enemy also has new quirk where they turn red when got shot.

this current scene is used for stress testing, so I have added a FPS counter on the top left of the screen that mactches with the phone resolution. The enemy spawner and the numbers of tower are just for stress testing.