Added tower spawning mechanic and camera movement
Added camera movement mechanic, with addition of clam function to prevent camera move more than it needs to also with threshold of 0.5f.

Added Spawning Tower mechanic where you can pressed the toggle on the bottom left of the screen and press anywhere on the screen to spawn a functional tower.