added event system, change tower behaviour, added basic Tweening, and change the Enumerator naming
Added event, so that an ui is activated when Koto Tower or Generator is off screen and diactivated when they are on screen. Change on the enumeration naming(now all UpperCase). Create a simple Tweening script where you can test it by clicking the first option on the answer Tab near "Koto Tower". Change behaviour on tower so it can only target enemy instead the "Char Charge"